Change store behaviour (read below)
Previously stores had an `onChange` that lets you define a callback that the store will call whenever you run API queries.
That means that pages might get re-rendered when inactive (the callback happens when you switch to a new page) causing the page to render over the current.
By moving to a Switch to manage Routes we prevent this, but now pages are unloaded and their observing methods are still held by the emitter.
This means going back and forth increases the number of duplicate callbacks to non-existing pages and causes massive memory leaks.
So they have been removed and now users should simply put the callbacks in the API query as an argument (e.g. `fetch(callback)`)